Split out a separate module for fake device servers
This will prevent generator.py from getting too loated.
